feat: Add community, meals, and meals share pages
This commit adds the following new pages:
- Community page
- Meals page
- Meals share page

Each page is a functional component that renders a heading. The pages are added to the existing `page.js` file.

The purpose of this commit is to provide the initial implementation of these pages, which will be further developed and styled in subsequent commits.