feat: Add getMeals function to retrieve meals from the database
The getMeals function is added to the meals.js file, allowing the retrieval of meals from the database. This function uses the better-sqlite3 library to connect to the "meals.db" database and executes a SELECT query to fetch all the meals. Additionally, a 2-second delay is added using a Promise to simulate a slow database operation. This commit enhances the functionality of the meals module.