feat: Add ImagePicker component for selecting and displaying images
The ImagePicker component is added to allow users to select and display images. This component includes an input field for selecting an image file and a button for triggering the file picker. This commit enhances the functionality of the application by providing a convenient way for users to pick images.